Level_3/Resume/requirements.txt
groq==0.32.0
PyPDF2==3.0.1
python-dotenv==1.1.1
streamlit==1.50.0
